+++ packages/expo-calendar/android/src/main/java/expo/modules/calendar/EventNotSavedException.java
@@ -1,4 +1,0 @@
-package expo.modules.calendar;
-
-public class EventNotSavedException extends Exception {
-}