+++ packages/expo-calendar/CHANGELOG.md
@@ -13,6 +13,8 @@
 
 ### 💡 Others
 
+- [Android] Refactored module native code. ([#40548](https://github.com/expo/expo/pull/40548) by [@barthap](https://github.com/barthap))
+
 ## 15.0.7 — 2025-09-11
 
 _This version does not introduce any user-facing changes._